requirements.txt
py_cpuinfo==7.0.0
tqdm==4.43.0
tabulate==0.8.7
PyYAML==5.3.1
Pillow==7.2.0

mathutils==2.81.2
numpy==1.18.1
scipy==1.4.1

torch==1.6.0
onnx==1.7.0
onnxruntime==1.5.1

bpy==2.82.1